Attempt to make harness more robust
The most important change here is that we add the ↵ test type, which
expects a CR or LF ahead of its text to match.  Ideally we'd assert
this every time, but there's a variety of extraneous stuff shells like
to print that can interfere with this.

I've also tweaked the timing a bit, trying to find a balance between
having the suite run quickly, while accommodating slower machines and
other sources of inconsistency.  I'm sure there will still be more
tweaks.

Fixes #10.